% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpimethexAnalysis.R
\name{epimethex.analysis}
\alias{epimethex.analysis}
\title{Create a Epimethex function}
\source{
Functions
}
\usage{
epimethex.analysis(dfExpressions, dfGpl, dfMethylation, minRangeGene,
    maxRangeGene, numCores, dataGenesLinear, testExpression, testMethylation)
}
\arguments{
\item{dfExpressions}{data frame, expression of data}

\item{dfGpl}{data frame, Annotations of data}

\item{dfMethylation}{data frame, Methylation of data}

\item{minRangeGene}{Numeric, lowerbound of genes}

\item{maxRangeGene}{Numeric, upperbound of genes}

\item{numCores}{Numeric, is the number of cores that you can use}

\item{dataGenesLinear}{Logic, determines if genetic data are linear}

\item{testExpression}{Logic, determines the test to apply on expression
dataset. If TRUE will apply t-student test,
otherwise will apply Kolmogorov-Smirnov test}

\item{testMethylation}{Logic, determines the test to apply on methylation
dataset. If TRUE will apply t-student test,
otherwise will apply Kolmogorov-Smirnov test}
}
\value{
csv files
}
\description{
Create a Epimethex function
}
\examples{

Annotations <- data.frame(
ID = c("cg11663302","cg01552731", "cg09081385"),
Relation_to_UCSC_CpG_Island = c("Island","N_Shore","N_Shore"),
UCSC_CpG_Islands_Name = c("chr1:18023481-18023792","chr19:46806998-46807617",
"chr12:120972167-120972447"),
UCSC_RefGene_Accession = c("NM_001011722","NM_152794","NM_014868"),
Chromosome_36 = c("1","19","12"),
Coordinate_36 = c("17896255","51498747","119456453"),
UCSC_RefGene_Name = c("ARHGEF10L","HIF3A","RNF10"),
UCSC_RefGene_Group =c("Body","1stExon","TSS200"),
stringsAsFactors=FALSE)

Expressions <- data.frame(
'sample' = c("ARHGEF10L", "HIF3A", "RNF10"),
'TCGA-YD-A89C-06' = c(-0.746592469762, -0.753826336325, 0.4953280),
'TCGA-Z2-AA3V-06' = c(0.578807530238, -2.30662633632, 0.1023280),
'TCGA-EB-A3Y6-01' = c(-0.363492469762, -2.67922633632, -0.6147720),
'TCGA-EE-A3JA-06' = c(-2.97279246976, -3.61932633632, 0.02932801),
'TCGA-D9-A4Z2-01' = c(-0.128492469762, 0.679073663675, 0.4017280),
'TCGA-D3-A51G-06' = c(-0.4299925, -4.0626263, -1.0136720),
stringsAsFactors=FALSE)

Methylation <- data.frame(
'sample' = c("cg11663302", "cg01552731", "cg09081385"),
'TCGA-YD-A89C-06' = c(0.9856, 0.7681, 0.0407),
'TCGA-Z2-AA3V-06' = c(0.9863, 0.8551, 0.0244),
'TCGA-EB-A3Y6-01' = c(0.9876, 0.6473, 0.028),
'TCGA-EE-A3JA-06' = c(0.9826, 0.4587, 0.0343),
'TCGA-D9-A4Z2-01' = c(0.9881, 0.8509, 0.0215),
'TCGA-D3-A51G-06' = c(0.9774, 0.813, 0.0332),
stringsAsFactors=FALSE)

epimethex.analysis(Expressions, Annotations, Methylation, 1, 5, 2,
TRUE, TRUE, FALSE)

}
