% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epimethex.R
\docType{package}
\name{Epimethex}
\alias{Epimethex}
\alias{Epimethex-package}
\title{Epimethex Package}
\description{
a tool for large-scale integrated analysis in methylation hotspots linked to genetic regulation
}
\author{
Giuseppe
}
