% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpimethexAnalysis.R
\name{epimethex.analysis}
\alias{epimethex.analysis}
\title{Create a Epimethex function}
\source{
Functions
}
\usage{
epimethex.analysis(path_dfPancan, path_dfGpl, path_dfMethylation, minRangeGene,
    maxRangeGene, numCores, dataLinear, outputPath)
}
\arguments{
\item{path_dfPancan}{String, is the path of expression data}

\item{path_dfGpl}{String, is the path of Annotations data}

\item{path_dfMethylation}{String, is the path of Methylation data}

\item{minRangeGene}{Numeric}

\item{maxRangeGene}{Numeric}

\item{numCores}{Numeric, is the number of cores that you can use}

\item{dataLinear}{Logic.}

\item{outputPath}{String}
}
\value{
files excel
}
\description{
Create a Epimethex function
}
